requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-decouple==2.1
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker-master==0.0.2
docker-py==1.10.6
docker-pycreds==0.4.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
psycopg2==2.9.5
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
syn==0.0.14
uritemplate==4.1.1
urllib3==1.26.14
websocket-client==1.4.2
